Move service functions into utils and update all imports
