+++ backend/app/api/routes.py
@@ -18,16 +18,11 @@ from datetime import datetime
 from stat import S_IWUSR, S_IREAD
 from dotenv import load_dotenv
 
-
-from services.fake_preprocess import Fake_preprocessor
 from database.database import Database
-from services.preprocess_bug_report import preprocess_bug_report
-from services.preprocess_source_code import preprocess_source_code
-from services.extract_gui_data import extract_gs_terms
-from services.extract_gui_data import extract_sc_terms
-from services.extract_gui_data import build_corpus
-from services.extract_gui_data import get_boosted_files
-from services.filter import filter_files
+from services.utils.preprocess_bug_report import preprocess_bug_report
+from services.utils.preprocess_source_code import preprocess_source_code
+from services.utils.extract_gui_data import extract_gs_terms, extract_sc_terms, build_corpus, get_boosted_files
+from services.utils.filter import filter_files
 from experimental_unixcoder.bug_localization import BugLocalization
 
 # Initialize Database
