+++ backend/services/utils/extract_gui_data.py
