+++ backend/tests/test_filter.py
@@ -1,4 +1,4 @@
-from services.filter import filter_files
+from services.utils.filter import filter_files
 import pytest
 from pathlib import Path
