+++ backend/tests/test_preprocess.py
@@ -1,5 +1,5 @@
 import pytest
-from services.preprocess import Preprocessor
+from services.utils.preprocess import Preprocessor
 from nltk.corpus import wordnet as wn
 
 def test_camel_case_split():
